app/requirements/dev.txt
-r base.txt

coverage==5.1
django-debug-toolbar==3.5.0
tblib==1.7.0
black
isort